Start modifying to decomposition
