+++ projectq/ops/__init__.py
@@ -36,4 +36,4 @@ from ._time_evolution import TimeEvolution
 from ._uniformly_controlled_rotation import (UniformlyControlledRy,
                                              UniformlyControlledRz)
 from ._state_prep import StatePreparation
-from ._phase_estimation import PhaseEstimation
+from ._qpegate import QPE
